WSL Localhost Trust failing:: Solution
In my case it was nothing more than telling Ubuntu to trust the Localhost certificate with the following command:
sudo update-ca-certificates

I did think that I was subject to constraints in this article:  https://www.fearofoblivion.com/setting-up-asp-net-dev-certs-for-wsl
I started to generate a localhost cert as defined in the article.  But again it was not needed.  The dev cert that we get with asp.net development does not have CA as an attribute of the Basic Constraints.  And as I mentioned it just needed to be trusted.  I am running Ubuntu 22.04.  I wanted to retain this branch I started incase in the future it may be needed.