Switch to Blazor Server only
The idea is that much of the work is x509 code which is not capable on the browser platform via WASM.  It turned out after pushing on this for a while it is very difficult to maintain the local browser storage.  I thought I could get rid of all the special model objects I created for serialization.  But then there seems to be a slow down when the local storage gets big, unlike using WASM.  Loosing connections between client and Blazor Server shows a retry and finally connecting.  But all in all it is a different kind of pain at the moment.  Going to put this on the shelf for a while.