Untrusted certificate validation bug fix
A previous change to the TrustChainValidator to enable termination at a chosen trust anchor broke the registration process for untrusted certificates.  Surprising there was not a test for this.  It appeared there was a test but it was only from the Server Metadata side and not the client registration.  The access token request would catch this.  Added a test.  Could put in more tests like when an anchor is removed etc...