Fixed a cancel registration bug
Found that an authorization_code flow style registration jwt used to cancel a registration exposed a bug that was introduced when the scope expansion work was put in.   Added another test to expose the issue.

Also found a logic error where I was setting the Duende Client model's RequirePkce all the time.  Changed it to respect the ServerSettings' RequirePkce.