Allow a Url style Uri to essentually ignore the trailing slash when comparing
So these two URIs would be considered the same:

https://dev-mtx-interop.meditech.com/
https://dev-mtx-interop.meditech.com

This shows up when encoding a sub-alt-name is a certificate as
https://dev-mtx-interop.meditech.com/
But the iss is set to https://dev-mtx-interop.meditech.com

With this change the two will still validate as the same.

I do need to produce my own certificate material for better long term tests.