Return the Set UdapStoreOptions in UdapDbContext constructor
Assigns UdapStoreOptions using the service provider in the UdapDbContext constructor to ensure configuration options are available during context initialization.

Turns out this is needed.  Not going to figure out how I can live without this for now.  There might be no way around this due to Duende.  But again just not solving that for now.  I will have to register a ConfigurationStoreOptions singleton to control the schema because of the way I create a inner ConfigurationDbContext from Duende.